src/requirements.txt
aenum==3.1.15
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
Deprecated==1.2.14
exceptiongroup==1.1.2
frozenlist==1.4.0
future==0.18.3
h11==0.14.0
idna==3.4
line-bot-sdk==3.1.0
multidict==6.0.4
outcome==1.2.0
pydantic==1.10.11
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.3
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
